- Made MediaMonitor make one instead of multiple "Mark all as played" folders - Made MediaMonitor constructor copy the passed array so it can't be modified from the caller - A couple of small bugfixes in MediaMonitor - A bugfix in DLNAResource

+++ src/main/java/net/pms/dlna/DLNAResource.java
@@ -2813,7 +2813,7 @@ public abstract class DLNAResource extends HTTPResource implements Cloneable, Ru
 	 *
 	 * @return The system time when the resource was last (re)started
 	 */
-	public double getLastStartSystemTime() {
+	public long getLastStartSystemTime() {
 		return lastStartSystemTime;
 	}
